Fix missing information in report
Address these comments:

The report doesn't state the correct number of data points.

The report should include a brief discussion to explain what parameter tuning is and why it is important. Also, explain what can happen if it isn't done well.

The definition of precision seems correct, but definition for recall seems to be explaining what precision is.